# [SPARK-13582] [SQL] defer dictionary decoding in parquet reader